requirements.txt
PyQt5
pillow
pillow-heif
